refactor: restructure query modules and fix default exports
