+++ src/hooks/mutations/todo/useCreateTodo.ts
@@ -4,7 +4,7 @@ import type { UseMutationCallback } from '@/types/types';
 import { todoKeys } from '@/constants/queryKeys';
 import { useSession } from '@/stores/session';
 
-export function useCreateTodo(callbacks?: UseMutationCallback) {
+export default function useCreateTodo(callbacks?: UseMutationCallback) {
   const queryClient = useQueryClient();
   const session = useSession();
   const userId = session?.user.id;
