+++ src/hooks/queries/todo/useAllTodos.ts
@@ -2,7 +2,7 @@ import { getAllTodos } from '@/api/todo';
 import { todoKeys } from '@/constants/queryKeys';
 import { useQuery } from '@tanstack/react-query';
 
-export function useAllTodos(userId?: string) {
+export default function useAllTodos(userId?: string) {
   return useQuery({
     queryKey: userId ? todoKeys.allTodos(userId) : [],
     queryFn: () => getAllTodos(userId!),
