+++ src/hooks/queries/todo/useHistoryTodos.ts
@@ -2,7 +2,10 @@ import { getTodosByDate } from '@/api/todo';
 import { todoKeys } from '@/constants/queryKeys';
 import { useQuery } from '@tanstack/react-query';
 
-export function useHistoryTodos(date: string, userId: string | undefined) {
+export default function useHistoryTodos(
+  date: string,
+  userId: string | undefined,
+) {
   return useQuery({
     queryKey: userId ? todoKeys.byDate(date, userId) : [],
     queryFn: () => getTodosByDate(date, userId!),
