+++ src/hooks/queries/todo/useTodoById.ts
@@ -2,7 +2,7 @@ import { getTodoById } from '@/api/todo';
 import { todoKeys } from '@/constants/queryKeys';
 import { useQuery } from '@tanstack/react-query';
 
-export function useTodoById(id?: string) {
+export default function useTodoById(id?: string) {
   return useQuery({
     queryKey: id ? todoKeys.byId(id) : [],
     queryFn: () => getTodoById(id!),
