+++ src/pages/FocusPage.tsx
@@ -2,10 +2,10 @@ import CustomDurationModal from '@/components/modal/CustomDurationModal';
 import FocusStopModal from '@/components/modal/FocusStopModal';
 import TimerCompletionModal from '@/components/modal/TimerCompletionModal';
 import { Button } from '@/components/ui/button';
-import { useCreateFocusSession } from '@/hooks/mutations/focus/useCreateFocusSession';
-import { useUpdateTodo } from '@/hooks/mutations/todo/useUpdateTodo';
+import useCreateFocusSession from '@/hooks/mutations/focus/useCreateFocusSession';
+import useUpdateTodo from '@/hooks/mutations/todo/useUpdateTodo';
 import useTodoFocusTime from '@/hooks/queries/focus/useTodoFocusTime';
-import { useTodoById } from '@/hooks/queries/useTodoById';
+import useTodoById from '@/hooks/queries/todo/useTodoById';
 import useTimer from '@/hooks/useTimer';
 import { formatTime } from '@/lib/utils';
 import { useSession } from '@/stores/session';
