+++ src/pages/TodayPage.tsx
@@ -12,7 +12,7 @@ import ReadOnlyMessage from '@/components/common/ReadOnlyMessage';
 import EmptyTodo from '@/components/common/EmptyTodo';
 import ErrorState from '@/components/common/ErrorState';
 import dayjs from 'dayjs';
-import { useTodayTodos } from '@/hooks/queries/useTodayTodos';
+import useTodayTodos from '@/hooks/queries/todo/useTodayTodos';
 import useTodayFocusTime from '@/hooks/queries/focus/useTodayFocusTime';
 
 type ReadOnlyVariant = 'past' | 'future';
